Integrated Petit FatFS lib with TI MMC lib

+++ filesys/pff.h
@@ -34,7 +34,8 @@ extern "C" {
 #include <windows.h>
 #elif (defined(__STDC_VERSION__) && __STDC_VERSION__ >= 199901L) || defined(__cplusplus)	/* C99 or later */
 #include <stdint.h>
-typedef unsigned int	UINT;	/* int must be 16-bit or 32-bit */
+// UINT changed to long for compatibility with MMC lib
+typedef unsigned long	UINT;	/* int must be 16-bit or 32-bit */
 typedef unsigned char	BYTE;	/* char must be 8-bit */
 typedef uint16_t		WORD;	/* 16-bit unsigned integer */
 typedef uint16_t		WCHAR;	/* 16-bit unsigned integer */
